Revert "Add prisma apis and change dataUpload component"
