+++ .env
@@ -1,15 +1,5 @@
-
-# Environment variables declared in this file are automatically made available to Prisma.
-# See the documentation for more detail: https://pris.ly/d/prisma-schema#using-environment-variables
-
-# Prisma supports the native connection string format for PostgreSQL, MySQL, SQLite, SQL Server (Preview) and MongoDB (Preview).
-# See the documentation for all the connection string options: https://pris.ly/d/connection-strings
-
-DATABASE_URL="file:./dev.db"
-
 AUTH0_SECRET='use [openssl rand -hex 32] to generate a 32 bytes value'
 AUTH0_BASE_URL='http://localhost:3000'
 AUTH0_ISSUER_BASE_URL='https://dev-w30xrd14.us.auth0.com'
 AUTH0_CLIENT_ID='QajeOio7F48oW781Q0jsz3VJ3rr45spx'
 AUTH0_CLIENT_SECRET='8-CnFbJr9IZ_0oPlmG6FB-cR4kls0WCiKaFMqud2aTPxCOzYBQDVGSdj4-wyAfhR'
-
