+++ .gitignore
@@ -1,4 +1,3 @@
-
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
@@ -12,7 +11,6 @@
 # next.js
 /.next/
 /out/
-.next
 
 # production
 /build
