+++ package.json
@@ -11,9 +11,7 @@
   "dependencies": {
     "@auth0/nextjs-auth0": "^1.5.0",
     "@emotion/css": "^11.1.3",
-    "@prisma/client": "^3.2.1",
     "bootstrap": "^5.1.0",
-    "formik": "^2.2.9",
     "next": "11.1.0",
     "react": "^17.0.2",
     "react-bootstrap": "^2.0.0-beta.6",
@@ -22,7 +20,6 @@
   },
   "devDependencies": {
     "eslint": "7.32.0",
-    "eslint-config-next": "11.1.0",
-    "prisma": "^3.2.1"
+    "eslint-config-next": "11.1.0"
   }
 }
