+++ pages/components/URLInput.js
@@ -1,6 +1,5 @@
 import React from "react";
 import { ReactDOM } from "react";
-import DBClient from '../../util/DBClient.js'
 
 class URLInput extends React.Component {
     constructor(props) {
@@ -11,7 +10,6 @@ class URLInput extends React.Component {
       event.preventDefault();
       alert(
         `URL - ${this.URLInput}`    );
-
     }
   
     render() {
@@ -31,24 +29,6 @@ class URLInput extends React.Component {
 
       );
     }
-
-    async createJob() {
-      const prisma = DBClient.getPrismaInstance()
-      const job = await prisma.job.create({
-            data: {
-                id: '20',
-                jobName : 'idk',
-                createTime : 'sdf',
-                status : 'sfsd',
-                inputLocation : 'sdf',
-                outputLocation : 'sdf',
-                author: '1',
-                authorID: '1',
-            }
-        })
-    }
-
   }
-
   
 export default URLInput;
