+++ pages/index.js
@@ -1,7 +1,6 @@
 import Head from 'next/head'
 import Image from 'next/image'
 import styles from '../styles/Home.module.css'
-import { Field, Formik, Form } from 'formik';
 
 import MainNavbar from './components/MainNavbar'
 import { useUser } from '@auth0/nextjs-auth0';
@@ -33,11 +32,10 @@ export default function Home() {
         </h1>
         <p className={styles.description}>
           Get started by inputting a{' '}
-          <code className={styles.code}>file or URL</code>
+          <code className={styles.code}>file</code>
           {/* . */}
         </p>
       </main>
-
       <footer className={styles.footer}>
         <a
           href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
@@ -51,6 +49,7 @@ export default function Home() {
         </a>
       </footer>
     </div>
-
   )
 }
+
+export { default as DataUpload } from "./components/DataUpload";
