+++ styles/Home.module.css
@@ -81,7 +81,7 @@
 .card {
   margin: 1rem;
   padding: 1.5rem;
-  text-align: center;
+  text-align: left;
   color: inherit;
   text-decoration: none;
   border: 1px solid #eaeaea;
@@ -93,21 +93,13 @@
 .card:hover,
 .card:focus,
 .card:active {
-  /* color: #0070f3; */
-  color: #000000;
-  border-color:#000000;
-  /* border-color: #0070f3; */
+  color: #0070f3;
+  border-color: #0070f3;
 }
 
 .card h2 {
   margin: 0 0 1rem 0;
   font-size: 1.5rem;
-}
-
-.card field {
-  margin: 0 0 1rem 0;
-  margin-left: 10%;
-  padding-left: 10%;
 }
 
 .card p {
@@ -123,11 +115,12 @@
 
 .uploadTitle {
   font-size: 1.25rem;
+
 }
 
 .dataUpload {
-  margin: 20;
-  padding: 20;
+  margin: 0;
+  padding: 0;
 
 }
 
@@ -138,12 +131,3 @@
   }
 }
 
-.dataUploadForm {
-  object-position: 50% 50%;
-  text-decoration-color: #0070f3;
-}
-
-.submitField {
-  margin: 20;
-  padding: 20;
-}
